# go-laos-geo/docs/openapi.json
{
    "swagger": "2.0",
    "info": {
        "description": "API to retrieve provinces, districts, and villages in Laos.",
        "title": "Laos Geo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/api/v1/districts/{id}/villages": {
            "get": {
                "description": "Get a paginated list of villages belonging to a specific district",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geo"
                ],
                "summary": "Get villages by district",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "District ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.StandardResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/provinces": {
            "get": {
                "description": "Get a paginated list of all provinces in Laos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geo"
                ],
                "summary": "Get all provinces",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.StandardResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/provinces/{id}/districts": {
            "get": {
                "description": "Get a paginated list of districts belonging to a specific province",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geo"
                ],
                "summary": "Get districts by province",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Province ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.StandardResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "current_page_total_item": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "domain.ResponseData": {
            "type": "object",
            "properties": {
                "list_data": {},
                "pagination": {
                    "$ref": "#/definitions/domain.Pagination"
                }
            }
        },
        "domain.StandardResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/domain.ResponseData"
                },
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}